# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "version": "1.0.0"
    },
    "paths": {
        "/api/ajout/create": {
            "post": {
                "summary": "Ajouter un produit",
                "operationId": "6e47f3ac152e1f34bec3161780deada7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "qte_ajouter",
                                    "product_id"
                                ],
                                "properties": {
                                    "qte_ajouter": {
                                        "description": "quantite ajouter",
                                        "type": "integer",
                                        "example": "30"
                                    },
                                    "product_id": {
                                        "description": "Identifiant du produit",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "produit ajouter  avec succès"
                    },
                    "422": {
                        "description": "Erreurs de validation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/product/create": {
            "post": {
                "summary": "Création d'un produit",
                "operationId": "8f4e29424e44f59b79251dc05369d58d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "designation",
                                    "description",
                                    "pa",
                                    "pv",
                                    "qte_alr",
                                    "qte_stock"
                                ],
                                "properties": {
                                    "designation": {
                                        "description": "Libelle du Produit",
                                        "type": "string",
                                        "example": "Souris Usb"
                                    },
                                    "description": {
                                        "description": "desription du produit",
                                        "type": "string",
                                        "example": "souris filaire"
                                    },
                                    "pa": {
                                        "description": "prix d'achat du produit",
                                        "type": "number",
                                        "format": "float",
                                        "example": 2500
                                    },
                                    "pv": {
                                        "description": "prix de vente du produit",
                                        "type": "number",
                                        "format": "float",
                                        "example": 5000
                                    },
                                    "qte_alr": {
                                        "description": "Quantite d'alerte du produit",
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "qte_stock": {
                                        "description": "Quantite en stock du produit",
                                        "type": "integer",
                                        "example": 10
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "produit  cree avec succes"
                    },
                    "422": {
                        "description": "Erreurs de validation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/product/details": {
            "get": {
                "summary": "Affichage des produits",
                "operationId": "81a62639060dacff47348b0c4a5f3277",
                "responses": {
                    "400": {
                        "description": "Listes des produits"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/product/search": {
            "get": {
                "summary": "Rechercher dun produit",
                "operationId": "d80b7efa5b94717d2aaeb1a38d9066cd",
                "parameters": [
                    {
                        "name": "designation",
                        "in": "query",
                        "description": "designation du produit",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/product/update": {
            "post": {
                "summary": "Mise a jour d'un produit",
                "operationId": "5b0774934c87184ce61a66f2a7851b3d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "designation",
                                    "description",
                                    "pa",
                                    "pv",
                                    "qte_alr",
                                    "qte_stock"
                                ],
                                "properties": {
                                    "designation": {
                                        "description": "Libelle du Produit",
                                        "type": "string",
                                        "example": "Souris Usb"
                                    },
                                    "description": {
                                        "description": "desription du produit",
                                        "type": "string",
                                        "example": "souris filaire"
                                    },
                                    "pa": {
                                        "description": "prix d'achat du produit",
                                        "type": "number",
                                        "format": "float",
                                        "example": 2500
                                    },
                                    "pv": {
                                        "description": "prix de vente du produit",
                                        "type": "number",
                                        "format": "float",
                                        "example": 5000
                                    },
                                    "qte_alr": {
                                        "description": "Quantite d'alerte du produit",
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "qte_stock": {
                                        "description": "Quantite en stock du produit",
                                        "type": "integer",
                                        "example": 10
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "produit  mis a jour avec succes"
                    },
                    "422": {
                        "description": "Erreurs de validation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/sells/create": {
            "post": {
                "summary": "Sortie d'un produit",
                "operationId": "c23a446d54b820b573a8c51ad3b41d1a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "product_id"
                                ],
                                "properties": {
                                    "product_id": {
                                        "description": "identifiant du produit",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "produit  sorti avec succes"
                    },
                    "422": {
                        "description": "Erreurs de validation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}